corrected colClasses used for geneList table import in HaemCode
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationHubData@83550 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/HaemCodeImportPreparer-class.R
@@ -134,11 +134,11 @@ setMethod("sourceUrls", "HaemCodeImportPreparer",
     entry <- 0
 
     geneList.read.table.colClasses <- c(rep("character", 6),
-                                        rep("integer",   3),
+                                        rep("numeric",   3),
                                         rep("character", 2),
-                                        rep("integer",   2),
+                                        rep("numeric",   2),
                                         rep("character", 1),
-                                        rep("integer",   2),
+                                        rep("numeric",   2),
                                         rep("character", 1))
     
     for(i in 1:max){
